修正

+++ db/sql/001_schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS `JmaXmlEntries` (
 # 震度速報、震源に関する情報、震源・震度に関する情報などは同じ地震に対してなので
 # これらはスレッドとして送信させる
 CREATE TABLE IF NOT EXISTS `TwitterThreads` (
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `tweet_id` VARCHAR(31) NOT NULL,
     `created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (`event_id`)
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS `TwitterThreads` (
 
 # 地震情報
 CREATE TABLE IF NOT EXISTS `Earthquakes` (
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `lat` INT,
     `lon` INT,
     `depth` INT,
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS `Earthquakes` (
 CREATE TABLE IF NOT EXISTS `TsunamiConnects` (
     `id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
     `tsunami_id` INT NOT NULL,
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (`id`)
 );
@@ -54,7 +54,7 @@ CREATE TABLE IF NOT EXISTS `TsunamiInfos` (
 # 震源・震度に関する情報
 CREATE TABLE IF NOT EXISTS `EarthquakeInfos` (
     `id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `lat` INT NOT NULL,
     `lon` INT NOT NULL,
     `depth` INT NOT NULL,
@@ -71,7 +71,7 @@ CREATE TABLE IF NOT EXISTS `EarthquakeInfos` (
 # 震源に関する情報
 CREATE TABLE IF NOT EXISTS `EarthquakeEpicenters` (
     `id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `lat` INT NOT NULL,
     `lon` INT NOT NULL,
     `depth` INT NOT NULL,
@@ -87,7 +87,7 @@ CREATE TABLE IF NOT EXISTS `EarthquakeEpicenters` (
 # 震度速報
 CREATE TABLE IF NOT EXISTS `EarthquakeReports` (
     `id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `max_int` VARCHAR(3) NOT NULL,
     `date` DATETIME NOT NULL,
     `created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -97,7 +97,7 @@ CREATE TABLE IF NOT EXISTS `EarthquakeReports` (
 
 CREATE TABLE IF NOT EXISTS `EarthquakeActivity` (
     `id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `date` DATETIME NOT NULL,
     `created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `row` TEXT NOT NULL,
@@ -106,7 +106,7 @@ CREATE TABLE IF NOT EXISTS `EarthquakeActivity` (
 
 CREATE TABLE IF NOT EXISTS `EarthquakeUpdate` (
     `id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `date` DATETIME NOT NULL,
     `created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `row` TEXT NOT NULL,
@@ -115,7 +115,7 @@ CREATE TABLE IF NOT EXISTS `EarthquakeUpdate` (
 
 CREATE TABLE IF NOT EXISTS `EarthquakeCount` (
     `id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
-    `event_id` INT NOT NULL,
+    `event_id` BIGINT NOT NULL,
     `date` DATETIME NOT NULL,
     `created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `row` TEXT NOT NULL,
