+++ src/models/EarthquakeEpicenters_test.go
@@ -568,7 +568,7 @@ func testEarthquakeEpicentersSelect(t *testing.T) {
 }
 
 var (
-	earthquakeEpicenterDBTypes = map[string]string{`ID`: `int`, `EventID`: `int`, `Lat`: `int`, `Lon`: `int`, `Depth`: `int`, `EpicenterName`: `text`, `Magnitude`: `int`, `MagnitudeType`: `varchar`, `Date`: `datetime`, `Created`: `datetime`, `Row`: `text`}
+	earthquakeEpicenterDBTypes = map[string]string{`ID`: `int`, `EventID`: `bigint`, `Lat`: `int`, `Lon`: `int`, `Depth`: `int`, `EpicenterName`: `text`, `Magnitude`: `int`, `MagnitudeType`: `varchar`, `Date`: `datetime`, `Created`: `datetime`, `Row`: `text`}
 	_                          = bytes.MinRead
 )
